I think I got the overflow voltage bug fixed with the drivetrain
it uses:
The code is hopefuly working the way I want it to. Because the motor groups can't take in a voltage range that is not within the bounds of -127 to 127 and because of the way that the voltage is currently calculated with our drivetrain system, we need to find a way to change the voltage of the other motor to counteract the system, not being able to give the motors of voltage to higher or low. We do this by finding the difference of voltage from 127 or -127 in one motor, and then taking that difference and adding it (or subtracting based on the needs of the specific case) to the other motor.